## Convert methods
As previously mentioned, any defined variables or methods will be exported by default. This allows us to pull individual methods into our Reason code and then import them into our JS.

Here, we pull the `responseBody` method into our Reason code. By using the `##body` syntax, Reason will infer that the argument to the function is of type `Js.t(< body: 'c, .. >)` which really just means that it must receive a JS object with the property `body` that has a value of any type.

We then import the `responseBody` method into our JS code and continue to use it the same way as before. If you haven't done so yet, now would be a good time to review the generated file `agent.bs.js` to see how all of this is compiling out. A goal of BuckleScript is to have readable output so hopefully you'll be able to mentally translate the output to your Reason code.