Moving `tokenPlugin` and `setToken` into our Reason code showcases optional values and mutations.
In Reason, we use "optional" values instead of `null` or `undefined` and the compiler will ensure we check whether we have `Some(data)` or `None`. When the application starts, we have `None` but what's this `ref()` thing? A `ref` is a language feature that allows us to create mutable variables.

We create the `token` variable as a mutable variable that starts as a `None` value.

Then, we need to rewrite the `tokenPlugin` a little bit. We change the if-statement to a `switch` that matches on `token^`. The caret after `token` means that we want to access the mutable value (it "unwraps" it from the `ref`). Our `switch` then compares the value to `Some(token)` or `None`, and if we have a token, it calls the `set` method on the `req` argument. If `token` is `None`, we need the statement `()` which is Reason's "non-value", usually called `unit` - this is because every code branch must return something and Reason has implicit returns.

Since we have a mutable value inside our Reason code, we also need to move the `setToken` method from our JS code into our Reason code. This method will be using the `:=` operator to set a mutable value. You'll also notice that we wrap the passed `_token` in `Some` to indicate that we now have some data assigned as `token`.

Once all that is done, you can import the `tokenPlugin` and `setToken` functions into your JS module and replace the JS implementations.