The superagent API is what people refer to as a "chaining" API - you've probably noticed the `.get().use().then()` syntax in the `agent.js` file.
We can represent these types of APIs in Reason using the `[@bs.send]` attribute combined with the Fast Pipe (`->`) operator. Fast Pipe inserts the results of a statement as the first argument to the next function in the chain. A "chaining" API will always return itself, which is `client` in our case.

For these externals, the empty string at the end indicates the method being called has the same name as our external. Note that `then` is a reserved word so we renamed it to `then_` and had to indicate the JS name is `"then"`.

In our `use` and `then_` externals, we've taken a shortcut by using generics (`'request` and `'response`) to indicate that the functions can receive "any" value.

The `then_` method will return a `Js.Promise.t(Js.Json.t)` value, which maps to a JS Promise containing some JSON value.

When we combine these externals with `->`, we get a Reason API that looks very similar to the JS chaining API.