### CommentList
We'll start by converting a pretty straight-forward, stateless component to Reason. Since our component is stateless, we'll construct our default component with `ReasonReact.statelessComponent()`. The only argument this method needs is a string to use as a debug name - a handy trick is to always use `__MODULE__` which Reason injects as the current filename (or module name if using nested modules).

In Reason, JSX compiles to the `make` method of a module, so we'll define a `make` that returns our ReasonReact component. All `props` are passed as labeled arguments, so here we define `~comments`, `~slug` and `~currentUser` - all of which are required (an optional property would look like `~token=?`). The only positional argument is `children` passed to the component, but in this case we define it as `_children` because we don't actually accept any children.

For the return value, we return a record that spreads the default `component` and then we override the `render` method. `render` receives one argument `self` - sort of like `this` in JS - but we aren't using it so we named it `_self`. This is a common pattern in Reason and reduces compiler warnings about unused variables.

Our `render` currently just outputs an empty `<div>` - but we'll be adding the rest from the original module next.