### ArticleMeta
Here's another fairly straightforward conversion. But we needed to change our `Link` component's `className` prop to an optional argument, which means we need to reach for the `Js.Nullable.fromOption` converter to change a `None` value into a `null` in our JS.

We also used a couple of Js.Date methods to work with a date string. Remember to wrap your strings in `ReasonReact.string`.

An extra `~onClickDelete` prop was added to this component because this application was over-abstracted and we need to thread these handlers throughout the components. If there's time later, we may show a pattern to tease apart these abstractions to make them nicer.