### CommentInput
We're finally getting into a `reducerComponent`! Before we begin, we need to define our `state` shape and the `action`s that can be taken. Remind you of anything? It's like a supercharged redux because we have a really powerful type system supporting it.

We have to define the `state` and `action` types before we call the `ReasonReact.reducerComponent` API. If we mess up the order, the compiler will let us know that they need to come first.