Next, we need a `reducer` method that receives an `action` taken and the current `state`. For this component, when we receive a `SetBody` action, we set the body to the payload of our action using the `ReasonReact.Update` API.
When we receive a `Submit` action, we use the `ReasonReact.UpdateWithSideEffects` API to reset the `body` property and trigger a side effect that creates the comment and calls the `onSubmit` prop. Notice that we send `state.body` in the API call, which was the previous body before the record was updated (because immutability!!!).

We also needed to duplicate another method from `agent.js`.