A couple of helper methods are needed so let's move our component down and add `setBody` and `createComment` functions at the top of our `make`. These both have the signature `(evt, self)` - the `self` is provided by the `self.handle()` we'll be using later.
Each of these functions utilize a couple of APIs - `ReactEvent.Form.*` for working with the events coming from our form elements. And `self.send()` to dispatch actions within this reducer component.

But why are we using `self.ReasonReact.send()`? Due to some indirection with these functions, we need to tell the type system where the `send()` method is coming from. This is a pattern you'll encounter a lot in ReasonReact.