### CommentContainer
Now that the dependencies of CommentContainer are all converted to Reason, we have a really easy time converting it too. One difference is that we change the `currentUser` prop to optional and then convert the if-statement to a switch that unwraps the optional value.

We'll also thread `onDelete` through the component.