### ListErrors (bonus bug fix!)
During the conversion we found a bug in the original code! Our CommentContainer component was using a `list-errors` element and we have no warning or indication this was an issue. Now that we found it, we can convert our ListErrors component to Reason and then use it in our CommentContainer.

A bonus is that BuckleScript has a `Js.Dict.entries()` API so we don't have to use the clunky `Object.keys()` technique like the original.

And even during the conversion of the ListErrors component, we found a visual bug that was caused by rendering an array of strings as a React element. Having a type system over this whole application would have caught that issue before it even happened. We've fixed the Reason version of ListErrors so it knows the shape of the errors prop.