### CommentList cleanup
Since we'll be using `CommentList` exclusively in Reason, we can clean up some of our earlier bindings and get better results from the type system. This also showed us that our quick-and-dirty bindings were missing quite a lot!

And our type system caught a bug! Our `comment##id` is an `int` but the type system inferred a string because we were combining it with another string.