### Article
Up until this point, we've been playing very fast-and-loose with our types, especially on objects coming from JS. Since we are now converting the `Article` component, we should get (at least a little) more serious about our types. These still aren't records, but at least we are defining the shape we are expecting to come back from our endpoints. Note: the better way to do this would be to decode a `Js.Json.t` type into concrete records but that exceeds our scope.