This component utilizes a `componentWillMount` hook in JS but ReasonReact doesn't support that lifecycle method so we use `didMount`.
We also had to use `Js.Promise.all2` because the 2 methods return different types of data - and every promise in `Js.Promise.all` must have matching types.

As mentioned previously, we're playing fast-and-loose with types so we had to relax the return value of our `agent.re` to a generic instead of `Js.Json.t`.