We've been threading methods throughout many of our components, and we're finally at the point where they are defined because they interact with the Article component's reducer.
Here we define `onCommentDelete` and `onCommentCreate` functions that wait for an API response and dispatch results into our reducer.

In Reason, `Js.Promise.catch` is a painful API to work with so we created a (very hacky) external that converts from some type to any other type using BuckleScripts "%identity". __Try your hardest not to use this and be extremely careful if you must.__ We inspected the shape of the object our `agent` APIs get rejected with and found that we can look up the response using this hack.